move etcd temp instance to its own module
